mova updateinfo to update.go
